+++ Model/Indexer/Product.php
@@ -34,7 +34,7 @@
  *
  */
 
-namespace Nosto\Tagging\Model\Indexer\Product;
+namespace Nosto\Tagging\Model\Indexer;
 
 use Magento\Catalog\Model\Product\Attribute\Source\Status;
 use Magento\Catalog\Model\Product\Visibility;
@@ -50,14 +50,15 @@ use Nosto\Exception\MemoryOutOfBoundsException;
 use Nosto\Tagging\Helper\Account as NostoHelperAccount;
 use Magento\Framework\Phrase;
 use Magento\Framework\Exception\LocalizedException;
+use Nosto\Tagging\Model\Product\Builder as NostoProductBuilder;
 
 /**
  * An indexer for Nosto product sync
  */
-class Indexer implements IndexerActionInterface, MviewActionInterface
+class Product implements IndexerActionInterface, MviewActionInterface
 {
     const BATCH_SIZE = 1000;
-    const INDEXER_ID = 'nosto_product_sync';
+    const INDEXER_ID = 'nosto_product_index';
 
     private $productService;
     private $dataHelper;
@@ -65,6 +66,7 @@ class Indexer implements IndexerActionInterface, MviewActionInterface
     private $productCollectionFactory;
     private $nostoQueueRepository;
     private $nostoHelperAccount;
+    private $nostoProductBuilder;
 
     /**
      * @param ProductService $productService
@@ -80,7 +82,8 @@ class Indexer implements IndexerActionInterface, MviewActionInterface
         NostoLogger $logger,
         ProductCollectionFactory $productCollectionFactory,
         NostoQueueRepository $nostoQueueRepository,
-        NostoHelperAccount\Proxy $nostoHelperAccount
+        NostoHelperAccount\Proxy $nostoHelperAccount,
+        NostoProductBuilder $nostoProductBuilder
     ) {
         $this->productService = $productService;
         $this->dataHelper = $dataHelper;
@@ -88,6 +91,7 @@ class Indexer implements IndexerActionInterface, MviewActionInterface
         $this->productCollectionFactory = $productCollectionFactory;
         $this->nostoQueueRepository = $nostoQueueRepository;
         $this->nostoHelperAccount = $nostoHelperAccount;
+        $this->nostoProductBuilder = $nostoProductBuilder;
     }
 
     /**
@@ -96,48 +100,39 @@ class Indexer implements IndexerActionInterface, MviewActionInterface
      */
     public function executeFull()
     {
-        if ($this->dataHelper->isFullReindexEnabled()
-            && !empty($this->nostoHelperAccount->getStoresWithNosto())
-        ) {
-            // Truncate queue table before first execution if they have leftover products
-            if ($this->nostoQueueRepository->isQueuePopulated()) {
-                $this->nostoQueueRepository->truncate();
+        $storesWithNosto = $this->nostoHelperAccount->getStoresWithNosto();
+        if (!empty($storesWithNosto)) {
+            foreach ($storesWithNosto as $store) {
+                $this->logger->info(sprintf('Starting to process store %s', $store->getCode()));
+                $productCollection = $this->getProductCollection();
+                $productCollection->setPageSize(self::BATCH_SIZE);
+                $lastPage = $productCollection->getLastPageNumber();
+                $pageNumber = 1;
+                do {
+                    $productCollection->setCurPage($pageNumber);
+                    $productCollection->addAttributeToSelect('id')
+                        ->addAttributeToFilter(
+                            'status',
+                            ['eq' => Status::STATUS_ENABLED]
+                        )->addAttributeToFilter(
+                            'visibility',
+                            ['neq' => Visibility::VISIBILITY_NOT_VISIBLE]
+                        );
+
+                    foreach ($productCollection->getItems() as $product) {
+                        $this->nostoProductBuilder->buildAndIndex($product, $store);
+                    }
+                    $pageNumber++;
+
+                    //Todo - remove the break
+                    if ($pageNumber > 1) {
+                        break;
+                    }
+                } while ($pageNumber <= $lastPage);
             }
-            $productCollection = $this->getProductCollection();
-            $productCollection->setPageSize(self::BATCH_SIZE);
-            $lastPage = $productCollection->getLastPageNumber();
-            $pageNumber = 1;
-            do {
-                $productCollection->setCurPage($pageNumber);
-                $productCollection->addAttributeToSelect('id')
-                    ->addAttributeToFilter(
-                        'status',
-                        ['eq'=> Status::STATUS_ENABLED]
-                    )->addAttributeToFilter(
-                        'visibility',
-                        ['neq'=> Visibility::VISIBILITY_NOT_VISIBLE]
-                    );
-                $products = [];
-                foreach ($productCollection->getItems() as $product) {
-                    $products[$product->getId()] = $product->getTypeId();
-                }
-                $productCollection->clear();
-                $this->logger->logWithMemoryConsumption(
-                    sprintf('Indexing from executeFull, remaining pages: %d', $lastPage - $pageNumber)
-                );
-                try {
-                    $this->productService->update($products);
-                } catch (MemoryOutOfBoundsException $e) {
-                    throw new LocalizedException(new Phrase($e->getMessage()));
-                }
-                $this->productService->processed = [];
-                $products = null;
-                $pageNumber++;
-            } while ($pageNumber <= $lastPage);
-            $productCollection = null;
         } else {
             $this->logger->info(
-                'Skip full reindex since full reindex is disabled or Nosto account is not connected into any store view'
+                'Nosto account is not connected into any store view'
             );
         }
     }
@@ -166,17 +161,7 @@ class Indexer implements IndexerActionInterface, MviewActionInterface
      */
     public function execute($ids)
     {
-        if (!empty($this->nostoHelperAccount->getStoresWithNosto())) {
-            $this->logger->logWithMemoryConsumption(sprintf('Got %d ids from CL', count($ids)));
-            $splitted = array_chunk(array_unique($ids), self::BATCH_SIZE);
-            foreach ($splitted as $batch) {
-                $this->productService->updateByIds($batch);
-            }
-        } else {
-            $this->logger->info(
-                'Nosto account is not connected into any store view. Nothing to index.'
-            );
-        }
+        // ToDo
     }
 
     /**