+++ Model/Product/Builder.php
@@ -46,6 +46,7 @@ use Magento\Review\Model\ReviewFactory;
 use Magento\Store\Model\Store;
 use Nosto\NostoException;
 use Nosto\Object\Product\Product as NostoProduct;
+use Nosto\Tagging\Api\Data\ProductIndexInterface;
 use Nosto\Tagging\Helper\Currency as CurrencyHelper;
 use Nosto\Tagging\Helper\Data as NostoHelperData;
 use Nosto\Tagging\Helper\Price as NostoPriceHelper;
@@ -60,6 +61,9 @@ use Nosto\Object\ModelFilter;
 use Nosto\Tagging\Model\Product\Variation\Collection as PriceVariationCollection;
 use Nosto\Tagging\Helper\Variation as NostoVariationHelper;
 use Nosto\Tagging\Helper\Ratings as NostoRating;
+use Nosto\Tagging\Model\Product\IndexRepository;
+use Nosto\Tagging\Model\Product\Index;
+use Nosto\Tagging\Util\Product as ProductUtil;
 
 class Builder
 {
@@ -88,6 +92,8 @@ class Builder
     private $categoryRepository;
     private $attributeSetRepository;
     private $nostoRatingHelper;
+    private $nostoProductIndexRepository;
+    private $nostoIndexFactory;
 
     /**
      * Builder constructor.
@@ -129,7 +135,9 @@ class Builder
         StockRegistryInterface $stockRegistry,
         PriceVariationCollection $priceVariationCollection,
         NostoVariationHelper $nostoVariationHelper,
-        NostoRating $nostoRatingHelper
+        NostoRating $nostoRatingHelper,
+        IndexRepository $nostoProductIndexRepository,
+        IndexFactory $nostoIndexFactory
     ) {
         $this->nostoDataHelper = $nostoHelperData;
         $this->nostoPriceHelper = $priceHelper;
@@ -153,6 +161,8 @@ class Builder
         $this->priceVariationCollection = $priceVariationCollection;
         $this->nostoVariationHelper = $nostoVariationHelper;
         $this->nostoRatingHelper = $nostoRatingHelper;
+        $this->nostoProductIndexRepository = $nostoProductIndexRepository;
+        $this->nostoIndexFactory = $nostoIndexFactory;
     }
 
     /**
@@ -174,9 +184,15 @@ class Builder
             'nosto_product_load_before',
             ['product' => $nostoProduct, 'magentoProduct' => $product, 'modelFilter' => $modelFilter]
         );
+
         if (!$modelFilter->isValid()) {
             return null;
         }
+
+        //ToDo
+        // - fetch from product index
+        // - separate builder for base data (the data that is saved in index table
+
         try {
             $nostoProduct->setUrl($this->urlBuilder->getUrlInStore($product, $store));
             $nostoProduct->setProductId((string)$product->getId());
@@ -189,21 +205,6 @@ class Builder
                     $store
                 ),
                 $store
-            );
-
-            $nostoProduct->setPrice($price);
-            $listPrice = $this->nostoCurrencyHelper->convertToTaggingPrice(
-                $this->nostoPriceHelper->getProductDisplayPrice(
-                    $product,
-                    $store
-                ),
-                $store
-            );
-            $nostoProduct->setListPrice($listPrice);
-            $nostoProduct->setPriceCurrencyCode(
-                $this->nostoCurrencyHelper->getTaggingCurrency(
-                    $store
-                )->getCode()
             );
 
             if ($this->nostoCurrencyHelper->exchangeRatesInUse($store)) {
@@ -280,6 +281,24 @@ class Builder
             if ($this->nostoDataHelper->isTagDatePublishedEnabled($store)) {
                 $nostoProduct->setDatePublished($product->getCreatedAt());
             }
+
+            // These will be always fetched from price index tables
+            $nostoProduct->setPrice($price);
+            $listPrice = $this->nostoCurrencyHelper->convertToTaggingPrice(
+                $this->nostoPriceHelper->getProductDisplayPrice(
+                    $product,
+                    $store
+                ),
+                $store
+            );
+            $nostoProduct->setListPrice($listPrice);
+            $nostoProduct->setPriceCurrencyCode(
+                $this->nostoCurrencyHelper->getTaggingCurrency(
+                    $store
+                )->getCode()
+            );
+
+
         } catch (\Exception $e) {
             $this->logger->exception($e);
         }
@@ -462,6 +481,42 @@ class Builder
         $nostoProduct = new NostoProduct();
         $nostoProduct->setProductId((string)$productId);
         $nostoProduct->setAvailability(ProductInterface::DISCONTINUED);
+
+        return $nostoProduct;
+    }
+
+    /**
+     * ToDo - this should be moved to somewhere else - perhaps some kind of index service
+     *
+     * @param Product $product
+     * @param Store $store
+     * @param string $nostoScope
+     * @return NostoProduct|null
+     * @throws \Exception
+     */
+    public function buildAndIndex(
+        Product $product,
+        Store $store,
+        $nostoScope = self::NOSTO_SCOPE_API
+    ) {
+        $nostoProduct = $this->build($product, $store, $nostoScope);
+        $productIndex = $this->nostoProductIndexRepository->getOneByProductAndStore($product, $store);
+        if ($productIndex instanceof ProductIndexInterface) {
+            if (ProductUtil::isEqual($productIndex->getNostoProduct(), $nostoProduct)) {
+                return $nostoProduct;
+            }
+        } else {
+            $productIndex = $this->nostoIndexFactory->create();
+            $productIndex->getCreatedAt(new \DateTime('now'));
+        }
+
+        $productIndex->setInSync(false);
+        $productIndex->setUpdatedAt(new \DateTime('now'));
+        $productIndex->setNostoProduct($nostoProduct);
+        $productIndex->setMagentoProduct($product);
+        $productIndex->setStore($store);
+
+        $this->nostoProductIndexRepository->save($productIndex);
 
         return $nostoProduct;
     }