+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
+### not released
+* Implement Nosto product index to speedup the Nosto product building, indexing and sending the data to Nosto
+
 ### 3.8.2
 * Fix redirect_uri for multi-store setup with different domains